infilect/requirements.txt
Django==2.0
djangorestframework==3.7.7
Pillow==5.4.1
PyJWT==1.5.3
pytz==2018.9
